hw12/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prometheus-client==0.19.0
psutil==5.9.6
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2
torch==2.1.1
transformers==4.35.2
python-multipart==0.0.6